feat: strengthen webnovel style defaults
