+++ server/src/services/styleEngine/defaultAntiAiRules.ts
@@ -1,0 +1,251 @@
+import type { DefaultAntiAiRuleDefinition } from "./defaultTypes";
+
+export const DEFAULT_ANTI_AI_RULES: DefaultAntiAiRuleDefinition[] = [
+  {
+    key: "forbid-explicit-psychology",
+    name: "禁止解释型心理描写",
+    type: "forbidden",
+    severity: "high",
+    description: "禁止直接使用“他感到”“他意识到”等句式解释人物心理。",
+    detectPatterns: ["他感到", "她感到", "他意识到", "她意识到", "他明白了", "她明白了"],
+    rewriteSuggestion: "把心理解释改成动作、语气、停顿、环境反应或结果。",
+    promptInstruction: "禁止直接解释人物心理，人物状态必须通过行为、动作、语气或环境反应体现。",
+    autoRewrite: true,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "forbid-ending-elevation",
+    name: "禁止段尾升华",
+    type: "forbidden",
+    severity: "high",
+    description: "禁止在段尾或收尾处用总结句升华主题。",
+    detectPatterns: ["生活就是", "命运总会", "归根结底", "说到底", "这就是"],
+    rewriteSuggestion: "删除升华句，回到具体动作、现场或悬而未决的处境。",
+    promptInstruction: "禁止段尾升华或总结主题，收尾必须落回具体情境。",
+    autoRewrite: true,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "forbid-theme-summary",
+    name: "禁止总结主题",
+    type: "forbidden",
+    severity: "high",
+    description: "禁止把段落写成总结中心思想或提炼人生道理。",
+    detectPatterns: ["这说明", "这意味着", "归根结底", "其实就是"],
+    rewriteSuggestion: "删掉主题总结，让信息通过事件和结果自然显现。",
+    promptInstruction: "不要总结主题，不要替读者提炼中心思想。",
+    autoRewrite: true,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "forbid-direct-preaching",
+    name: "禁止直接说教",
+    type: "forbidden",
+    severity: "medium",
+    description: "禁止作者替角色或读者做直接价值判断和说教。",
+    detectPatterns: ["我们都应该", "人总要学会", "真正重要的是"],
+    rewriteSuggestion: "改成角色具体处境或对话，不做抽象说教。",
+    promptInstruction: "禁止直接说教或价值判断，让意义落在事件本身。",
+    autoRewrite: true,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "forbid-template-transitions",
+    name: "禁止模板化转场和导游词",
+    type: "forbidden",
+    severity: "high",
+    description: "禁止用“这一刻”“与此同时”“接下来”等提示词式转场或结论句推动章节。",
+    detectPatterns: ["这一刻", "与此同时", "接下来", "他知道，", "她知道，", "毫无疑问", "不难看出"],
+    rewriteSuggestion: "改用角色动作、视线变化、声音、物件或一句有信息量的对话完成场景切换。",
+    promptInstruction: "禁止写作提示词式转场、结论式提示和段首导游词；转场必须落在动作、视线、声音、物件或对话上。",
+    autoRewrite: true,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "forbid-generic-metaphor",
+    name: "禁止通用 AI 腔比喻和泛化形容",
+    type: "forbidden",
+    severity: "high",
+    description: "禁止使用“仿佛”“难以言喻”“复杂情绪”等可套进任何场景的泛化表达。",
+    detectPatterns: ["仿佛", "像是某种", "空气仿佛凝固", "命运的齿轮", "某种意义上", "无形之中", "复杂的情绪", "难以言喻"],
+    rewriteSuggestion: "替换为当前角色、地点、物件和冲突独有的具体细节。",
+    promptInstruction: "禁止通用 AI 腔套话和泛化形容词；必须换成当前角色或场景独有的具体细节。",
+    autoRewrite: true,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "forbid-setting-info-dump",
+    name: "禁止说明书式设定倾倒",
+    type: "forbidden",
+    severity: "high",
+    description: "禁止连续整段介绍世界规则、修炼体系、组织结构或背景资料。",
+    detectPatterns: ["这个世界", "所谓", "规则是", "体系分为", "简单来说", "众所周知"],
+    rewriteSuggestion: "把设定拆进冲突、选择、代价、误会、交易或失败后果里释放。",
+    promptInstruction: "设定只能在冲突、选择、代价、对话误解或后果中释放，不得写成说明书。",
+    autoRewrite: true,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "risk-even-paragraph-length",
+    name: "段落长度过于整齐",
+    type: "risk",
+    severity: "medium",
+    description: "段落长度和节奏过于平均，容易产生 AI 作文感。",
+    detectPatterns: ["段落整齐", "节奏平均"],
+    rewriteSuggestion: "打破段落长度均衡，让句子和段落有自然起伏。",
+    promptInstruction: "注意避免每段都过于工整、平均和像标准作文。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "risk-three-paragraphs-exposition",
+    name: "连续三段解释性叙事",
+    type: "risk",
+    severity: "high",
+    description: "连续几段只有解释没有动作，会削弱现场感。",
+    detectPatterns: ["连续解释", "没有动作"],
+    rewriteSuggestion: "插入动作、对话、环境反馈，减少连段说明。",
+    promptInstruction: "注意避免连续几段只有解释没有动作或对话。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "risk-dialogue-too-functional",
+    name: "对话纯功能推进",
+    type: "risk",
+    severity: "medium",
+    description: "对话只有信息推进，没有人物语气和生活噪音。",
+    detectPatterns: ["告诉你", "我们现在要", "接下来就"],
+    rewriteSuggestion: "补入停顿、绕弯、语气差异和无效信息。",
+    promptInstruction: "对话不要只承担推进剧情功能，要保留人物语气和生活噪音。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "risk-repeated-sentence-structure",
+    name: "句式重复率过高",
+    type: "risk",
+    severity: "medium",
+    description: "连续句式过于整齐，容易显得机械。",
+    detectPatterns: ["首先", "然后", "接着", "最后"],
+    rewriteSuggestion: "拉开句式长度和起句方式，打散结构。",
+    promptInstruction: "警惕连续句式重复和排比化表达。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "risk-scene-without-state-change",
+    name: "场景没有状态变化",
+    type: "risk",
+    severity: "high",
+    description: "场景只是在聊天、解释或铺垫，没有改变信息、关系、风险、资源或决定。",
+    detectPatterns: ["继续交谈", "只是铺垫", "保持不变", "没有变化"],
+    rewriteSuggestion: "给场景补一个可见变化：信息差、关系裂口、风险升级、资源损失、选择代价或新义务。",
+    promptInstruction: "每个重要场景必须带来可见状态变化；不能只有聊天、解释、氛围或铺垫。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: true,
+  },
+  {
+    key: "encourage-useless-action",
+    name: "鼓励无意义小动作",
+    type: "encourage",
+    severity: "low",
+    description: "鼓励加入真实但不推动主线的小动作增强人味。",
+    detectPatterns: [],
+    rewriteSuggestion: "补入挠头、点烟、抠包装、挪椅子等小动作。",
+    promptInstruction: "优先加入真实但不推动主线的小动作，增加生活感。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: false,
+  },
+  {
+    key: "encourage-reality-gap",
+    name: "鼓励现实落差",
+    type: "encourage",
+    severity: "low",
+    description: "让人物预期和现实结果之间出现落差。",
+    detectPatterns: [],
+    rewriteSuggestion: "补出人物预期与实际结果之间的差距。",
+    promptInstruction: "优先制造现实落差，让人物预期和结果不完全一致。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: false,
+  },
+  {
+    key: "encourage-hard-mouth-compensation",
+    name: "鼓励嘴硬补偿",
+    type: "encourage",
+    severity: "low",
+    description: "人物吃瘪后用嘴硬、转移或找补来维持体面。",
+    detectPatterns: [],
+    rewriteSuggestion: "给角色补一句嘴硬找补或自我合理化。",
+    promptInstruction: "人物吃瘪后可用嘴硬、转移或找补来补偿体面。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: false,
+  },
+  {
+    key: "encourage-life-noise",
+    name: "鼓励生活噪音",
+    type: "encourage",
+    severity: "low",
+    description: "适当加入与主线不完全相关的生活噪音增加现场感。",
+    detectPatterns: [],
+    rewriteSuggestion: "补入环境声、杂音、小打断或生活性信息。",
+    promptInstruction: "适当保留生活噪音和无效信息，增强现场感。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: false,
+  },
+  {
+    key: "encourage-character-specific-voice",
+    name: "鼓励角色专属口吻",
+    type: "encourage",
+    severity: "low",
+    description: "重要角色需要有可辨认的说话节奏、避讳、误解方式和小动作。",
+    detectPatterns: [],
+    rewriteSuggestion: "为每个重要角色补入口头节奏、避讳词、习惯动作或只属于他的误读方式。",
+    promptInstruction: "重要角色必须有可识别的口头节奏、避讳、误解方式和小动作；对话不能像同一个旁白代言。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: false,
+  },
+  {
+    key: "encourage-subtext-dialogue",
+    name: "鼓励对白潜台词",
+    type: "encourage",
+    severity: "low",
+    description: "对白要允许角色绕开真正想说的话，通过误读、试探、停顿或嘴硬暴露关系。",
+    detectPatterns: [],
+    rewriteSuggestion: "把直接说明改成试探、打岔、反问、沉默或不合时宜的小动作。",
+    promptInstruction: "重要对白要有潜台词或错位：角色不必直接说出真实意图，可以通过试探、停顿、嘴硬或误读暴露关系。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: false,
+  },
+  {
+    key: "encourage-scene-cost",
+    name: "鼓励场景代价",
+    type: "encourage",
+    severity: "low",
+    description: "场景推进后应留下代价、风险、义务或关系裂口，而不是只给主角好处。",
+    detectPatterns: [],
+    rewriteSuggestion: "在兑现收益后补一个代价：被人看见、欠下人情、暴露弱点、失去资源或推高风险。",
+    promptInstruction: "重要推进要带代价或后果；主角得到信息、资源或胜利时，最好同时留下新风险、义务或关系裂口。",
+    autoRewrite: false,
+    enabled: true,
+    globalBaselineEnabled: false,
+  },
+];
