+++ server/src/services/styleEngine/defaultStyleTemplates.ts
@@ -1,0 +1,435 @@
+import type { DefaultTemplateDefinition } from "./defaultTypes";
+
+export const DEFAULT_STYLE_TEMPLATES: DefaultTemplateDefinition[] = [
+  {
+    key: "bottom-loop-reality",
+    name: "底层循环现实流",
+    description: "通过碎片化生活与反复落空表现人物困境。",
+    category: "现实流",
+    tags: ["第一人称", "口语化", "碎片叙事"],
+    applicableGenres: ["都市", "现实", "成长"],
+    analysisMarkdown: "以时间推进和现实落差构成叙事张力，结尾不解决核心困境。",
+    narrativeRules: {
+      progressionMode: "time_sequence",
+      sceneUnitPattern: ["行为", "落差", "自我合理化"],
+      multiPov: false,
+      looping: true,
+      endingStyle: "unresolved",
+      summary: "以碎片化生活推进，不做总括式回顾。",
+    },
+    characterRules: {
+      allowSelfReflection: false,
+      emotionExpression: "behavior_only",
+      defenseMechanisms: ["嘴硬", "转移", "自我合理化"],
+      facePriority: true,
+      dialogueStyle: "short_colloquial",
+      summary: "人物情绪通过动作和嘴硬表达。",
+    },
+    languageRules: {
+      register: "colloquial",
+      roughness: 0.8,
+      allowIncompleteSentences: true,
+      allowSwearing: true,
+      sentenceVariation: "high",
+      allowUselessDetails: true,
+      summary: "语言粗粝、口语化，允许生活杂音。",
+    },
+    rhythmRules: {
+      pace: "medium_fast",
+      paragraphDensity: "high",
+      allowFragmentedFlow: true,
+      actionOverExplanation: true,
+      summary: "段落密实，动作先于解释。",
+    },
+    defaultAntiAiRuleKeys: [
+      "forbid-explicit-psychology",
+      "forbid-ending-elevation",
+      "encourage-useless-action",
+      "encourage-reality-gap",
+      "encourage-hard-mouth-compensation",
+    ],
+  },
+  {
+    key: "power-up-escalation",
+    name: "爽文递进推进流",
+    description: "持续升级冲突和收益点，强化目标推进与爽点兑现。",
+    category: "爽文流",
+    tags: ["推进感", "收益点", "冲突升级"],
+    applicableGenres: ["都市", "玄幻", "热血"],
+    analysisMarkdown: "每段都要有目标推进或爽点兑现，保持明确因果和节奏抬升。",
+    narrativeRules: {
+      progressionMode: "goal_driven",
+      sceneUnitPattern: ["目标", "阻碍", "压制", "反转收益"],
+      multiPov: false,
+      looping: false,
+      endingStyle: "hook",
+      summary: "围绕目标推进，尽快兑现局部收益。",
+    },
+    characterRules: {
+      allowSelfReflection: true,
+      emotionExpression: "dialogue_and_action",
+      defenseMechanisms: [],
+      facePriority: false,
+      dialogueStyle: "direct",
+      summary: "角色表达直接，情绪跟随胜负切换。",
+    },
+    languageRules: {
+      register: "direct",
+      roughness: 0.55,
+      allowIncompleteSentences: false,
+      allowSwearing: false,
+      sentenceVariation: "medium",
+      allowUselessDetails: false,
+      summary: "句式清晰，减少无效分散信息。",
+    },
+    rhythmRules: {
+      pace: "fast",
+      paragraphDensity: "medium",
+      allowFragmentedFlow: false,
+      actionOverExplanation: true,
+      summary: "优先冲突和结果，少停留。",
+    },
+    defaultAntiAiRuleKeys: [
+      "forbid-theme-summary",
+      "risk-dialogue-too-functional",
+      "risk-three-paragraphs-exposition",
+    ],
+  },
+  {
+    key: "suspense-pressure",
+    name: "悬疑压迫递增流",
+    description: "通过信息遮蔽、细节异常和压力叠加制造不安感。",
+    category: "悬疑流",
+    tags: ["压迫感", "信息差", "异常细节"],
+    applicableGenres: ["悬疑", "惊悚", "现实"],
+    analysisMarkdown: "以异常细节、信息差和节奏收束推动悬念层层加压。",
+    narrativeRules: {
+      progressionMode: "mystery_escalation",
+      sceneUnitPattern: ["现场细节", "异常", "误判", "新风险"],
+      multiPov: false,
+      looping: false,
+      endingStyle: "suspense",
+      summary: "优先制造信息缺口和压迫氛围。",
+    },
+    characterRules: {
+      allowSelfReflection: true,
+      emotionExpression: "reaction_only",
+      defenseMechanisms: ["压抑"],
+      facePriority: false,
+      dialogueStyle: "restrained",
+      summary: "角色反应克制，恐惧通过反应显现。",
+    },
+    languageRules: {
+      register: "restrained",
+      roughness: 0.45,
+      allowIncompleteSentences: true,
+      allowSwearing: false,
+      sentenceVariation: "medium_high",
+      allowUselessDetails: true,
+      summary: "细节精确，保留少量噪音增强现场感。",
+    },
+    rhythmRules: {
+      pace: "medium",
+      paragraphDensity: "medium_high",
+      allowFragmentedFlow: true,
+      actionOverExplanation: true,
+      summary: "通过节奏收束和信息延迟制造压力。",
+    },
+    defaultAntiAiRuleKeys: [
+      "forbid-explicit-psychology",
+      "forbid-theme-summary",
+      "risk-even-paragraph-length",
+      "encourage-reality-gap",
+    ],
+  },
+  {
+    key: "emotional-tension",
+    name: "情绪拉扯流",
+    description: "通过错位表达、停顿和误读制造关系张力。",
+    category: "情感流",
+    tags: ["误读", "拉扯", "停顿感"],
+    applicableGenres: ["言情", "都市", "群像"],
+    analysisMarkdown: "人物不直说核心情绪，靠误读、停顿和反应推动关系变化。",
+    narrativeRules: {
+      progressionMode: "relationship_push_pull",
+      sceneUnitPattern: ["动作", "言外之意", "误读", "回避"],
+      multiPov: false,
+      looping: false,
+      endingStyle: "emotional_hook",
+      summary: "以关系错位推进，而非直接说明。",
+    },
+    characterRules: {
+      allowSelfReflection: true,
+      emotionExpression: "subtext",
+      defenseMechanisms: ["回避", "试探", "嘴硬"],
+      facePriority: true,
+      dialogueStyle: "subtext_heavy",
+      summary: "情绪通过停顿、动作和言外之意体现。",
+    },
+    languageRules: {
+      register: "natural",
+      roughness: 0.35,
+      allowIncompleteSentences: true,
+      allowSwearing: false,
+      sentenceVariation: "high",
+      allowUselessDetails: true,
+      summary: "语言自然，允许留白与停顿。",
+    },
+    rhythmRules: {
+      pace: "medium_slow",
+      paragraphDensity: "medium",
+      allowFragmentedFlow: true,
+      actionOverExplanation: false,
+      summary: "给关系反应留空间，但避免空洞抒情。",
+    },
+    defaultAntiAiRuleKeys: [
+      "forbid-direct-preaching",
+      "forbid-ending-elevation",
+      "risk-dialogue-too-functional",
+      "encourage-useless-action",
+    ],
+  },
+  {
+    key: "ensemble-weave",
+    name: "群像交织流",
+    description: "以多人行动线和视角差异交织推进事件。",
+    category: "群像流",
+    tags: ["多角色", "交织", "信息流动"],
+    applicableGenres: ["群像", "都市", "悬疑"],
+    analysisMarkdown: "多角色并行推进，但每个角色的表达和认知范围必须区分清楚。",
+    narrativeRules: {
+      progressionMode: "multi_thread",
+      sceneUnitPattern: ["角色动作", "局部信息", "交叉影响"],
+      multiPov: true,
+      looping: false,
+      endingStyle: "cross_hook",
+      povSwitchStyle: "controlled",
+      summary: "多线并进，但视角切换要受控。",
+    },
+    characterRules: {
+      allowSelfReflection: true,
+      emotionExpression: "mixed",
+      defenseMechanisms: [],
+      facePriority: false,
+      dialogueStyle: "distinct_by_role",
+      summary: "不同角色口吻必须拉开差异。",
+    },
+    languageRules: {
+      register: "flexible",
+      roughness: 0.45,
+      allowIncompleteSentences: true,
+      allowSwearing: false,
+      sentenceVariation: "high",
+      allowUselessDetails: false,
+      summary: "保持角色差异，避免所有人说话一样。",
+    },
+    rhythmRules: {
+      pace: "balanced",
+      paragraphDensity: "medium",
+      allowFragmentedFlow: false,
+      actionOverExplanation: true,
+      summary: "多线交织但节奏不乱。",
+    },
+    defaultAntiAiRuleKeys: [
+      "risk-dialogue-too-functional",
+      "risk-repeated-sentence-structure",
+      "forbid-theme-summary",
+    ],
+  },
+  {
+    key: "immersive-daily",
+    name: "日常浸没流",
+    description: "通过生活细节和细微情绪变化建立持续沉浸感。",
+    category: "日常流",
+    tags: ["生活感", "沉浸", "细碎细节"],
+    applicableGenres: ["日常", "治愈", "都市"],
+    analysisMarkdown: "允许保留生活性动作和无效信息，但核心情绪仍要通过场景自然流出。",
+    narrativeRules: {
+      progressionMode: "scene_immersion",
+      sceneUnitPattern: ["动作", "环境", "关系反应"],
+      multiPov: false,
+      looping: false,
+      endingStyle: "soft_open",
+      summary: "重场景体验和关系温度。",
+    },
+    characterRules: {
+      allowSelfReflection: true,
+      emotionExpression: "light_behavior",
+      defenseMechanisms: [],
+      facePriority: false,
+      dialogueStyle: "daily_natural",
+      summary: "人物表达自然，不用高强度戏剧句。",
+    },
+    languageRules: {
+      register: "colloquial",
+      roughness: 0.25,
+      allowIncompleteSentences: true,
+      allowSwearing: false,
+      sentenceVariation: "medium_high",
+      allowUselessDetails: true,
+      summary: "保留生活细节和杂音，不追求工整。",
+    },
+    rhythmRules: {
+      pace: "slow",
+      paragraphDensity: "medium",
+      allowFragmentedFlow: true,
+      actionOverExplanation: false,
+      summary: "慢节奏沉浸，但避免空转。",
+    },
+    defaultAntiAiRuleKeys: [
+      "forbid-ending-elevation",
+      "risk-even-paragraph-length",
+      "encourage-useless-action",
+    ],
+  },
+  {
+    key: "cold-professional",
+    name: "冷峻专业流",
+    description: "以专业事实和行业细节压住情绪，形成克制压力感。",
+    category: "专业流",
+    tags: ["专业细节", "克制", "事实压情绪"],
+    applicableGenres: ["职场", "现实", "悬疑"],
+    analysisMarkdown: "行业事实和程序细节优先，情绪不直说，信息密度高于抒情密度。",
+    narrativeRules: {
+      progressionMode: "fact_driven",
+      sceneUnitPattern: ["事实", "动作", "专业判断", "后果"],
+      multiPov: false,
+      looping: false,
+      endingStyle: "pressure_continue",
+      summary: "让专业事实承担叙事重量。",
+    },
+    characterRules: {
+      allowSelfReflection: false,
+      emotionExpression: "suppressed",
+      defenseMechanisms: ["克制"],
+      facePriority: false,
+      dialogueStyle: "informational",
+      summary: "情绪藏在专业动作和事实选择里。",
+    },
+    languageRules: {
+      register: "professional",
+      roughness: 0.2,
+      allowIncompleteSentences: false,
+      allowSwearing: false,
+      sentenceVariation: "medium",
+      allowUselessDetails: false,
+      summary: "术语和事实优先，避免廉价金句。",
+    },
+    rhythmRules: {
+      pace: "balanced",
+      paragraphDensity: "medium_high",
+      allowFragmentedFlow: false,
+      actionOverExplanation: true,
+      summary: "信息密度高，但不铺张解释。",
+    },
+    defaultAntiAiRuleKeys: [
+      "forbid-direct-preaching",
+      "forbid-theme-summary",
+      "risk-repeated-sentence-structure",
+    ],
+  },
+  {
+    key: "absurd-dark-humor",
+    name: "荒诞黑色幽默流",
+    description: "通过反差、冷感观察和荒诞细节制造黑色幽默。",
+    category: "黑色幽默",
+    tags: ["荒诞", "反差", "冷感"],
+    applicableGenres: ["都市", "黑色幽默", "现实"],
+    analysisMarkdown: "用反差和荒诞细节放大现实困境，笑点和压迫感同时存在。",
+    narrativeRules: {
+      progressionMode: "contrast_driven",
+      sceneUnitPattern: ["现实细节", "荒诞偏差", "冷反应"],
+      multiPov: false,
+      looping: false,
+      endingStyle: "bitter_aftertaste",
+      summary: "依赖反差和冷感观察，而非热闹吐槽。",
+    },
+    characterRules: {
+      allowSelfReflection: false,
+      emotionExpression: "deadpan",
+      defenseMechanisms: ["自嘲", "转移"],
+      facePriority: true,
+      dialogueStyle: "deadpan_colloquial",
+      summary: "情绪藏在冷反应和嘴硬里。",
+    },
+    languageRules: {
+      register: "colloquial",
+      roughness: 0.5,
+      allowIncompleteSentences: true,
+      allowSwearing: true,
+      sentenceVariation: "high",
+      allowUselessDetails: true,
+      summary: "允许夹带荒诞杂质和冷幽默节奏。",
+    },
+    rhythmRules: {
+      pace: "balanced",
+      paragraphDensity: "medium_high",
+      allowFragmentedFlow: true,
+      actionOverExplanation: true,
+      summary: "反差点要快落地，不要解释笑点。",
+    },
+    defaultAntiAiRuleKeys: [
+      "forbid-explicit-psychology",
+      "forbid-ending-elevation",
+      "encourage-reality-gap",
+      "encourage-hard-mouth-compensation",
+    ],
+  },
+  {
+    key: "webnovel-human-continuity",
+    name: "网文自然人味连续流",
+    description: "默认给长篇网文使用：按写手级四轮过闸，强化现场推进、人物口吻、场景代价和连续性。",
+    category: "长篇网文",
+    tags: ["去AI味", "连续性", "人味", "网文"],
+    applicableGenres: ["玄幻", "都市", "悬疑", "科幻", "言情", "历史"],
+    analysisMarkdown: "先按优秀写手的判断方式看每章是否有场景功能、角色选择、代价变化和连续性承接，再处理去 AI 味。正文必须像真实现场，而不是像模型在总结设定。",
+    narrativeRules: {
+      progressionMode: "chapter_chain",
+      sceneUnitPattern: ["场景功能", "当前压力", "角色选择", "代价变化", "章末钩子"],
+      multiPov: false,
+      looping: false,
+      endingStyle: "hook",
+      summary: "每章只解决局部问题，同时带来新的代价、压力、选择或误解。",
+    },
+    characterRules: {
+      allowSelfReflection: false,
+      emotionExpression: "dialogue_and_action",
+      defenseMechanisms: ["嘴硬", "试探", "遮掩", "找补"],
+      facePriority: true,
+      dialogueStyle: "distinct_by_role",
+      summary: "人物通过动作、口头禅、回避、潜台词和选择暴露状态，不替他们解释。",
+    },
+    languageRules: {
+      register: "colloquial",
+      roughness: 0.55,
+      allowIncompleteSentences: true,
+      allowSwearing: false,
+      sentenceVariation: "high",
+      allowUselessDetails: true,
+      summary: "语言要像人在现场说话和观察，保留不工整的短句、停顿和生活杂音。",
+    },
+    rhythmRules: {
+      pace: "medium_fast",
+      paragraphDensity: "medium_high",
+      allowFragmentedFlow: true,
+      actionOverExplanation: true,
+      summary: "先让事情发生，再让读者从后果里理解。",
+    },
+    defaultAntiAiRuleKeys: [
+      "forbid-template-transitions",
+      "forbid-generic-metaphor",
+      "forbid-setting-info-dump",
+      "forbid-explicit-psychology",
+      "forbid-theme-summary",
+      "risk-dialogue-too-functional",
+      "risk-scene-without-state-change",
+      "risk-repeated-sentence-structure",
+      "risk-even-paragraph-length",
+      "encourage-useless-action",
+      "encourage-life-noise",
+      "encourage-character-specific-voice",
+      "encourage-subtext-dialogue",
+      "encourage-scene-cost",
+    ],
+  },
+];
