+++ server/src/services/styleEngine/defaults.ts
@@ -1,719 +1,8 @@
-import type {
-  AntiAiRuleType,
-  AntiAiSeverity,
-  CharacterRules,
-  LanguageRules,
-  NarrativeRules,
-  RhythmRules,
-} from "@ai-novel/shared/types/styleEngine";
-
-export interface DefaultTemplateDefinition {
-  key: string;
-  name: string;
-  description: string;
-  category: string;
-  tags: string[];
-  applicableGenres: string[];
-  analysisMarkdown: string;
-  narrativeRules: NarrativeRules;
-  characterRules: CharacterRules;
-  languageRules: LanguageRules;
-  rhythmRules: RhythmRules;
-  defaultAntiAiRuleKeys: string[];
-}
-
-export interface DefaultAntiAiRuleDefinition {
-  key: string;
-  name: string;
-  type: AntiAiRuleType;
-  severity: AntiAiSeverity;
-  description: string;
-  detectPatterns: string[];
-  rewriteSuggestion: string;
-  promptInstruction: string;
-  autoRewrite: boolean;
-  enabled: boolean;
-  globalBaselineEnabled: boolean;
-}
-
-export interface DefaultStarterStyleProfileDefinition {
-  key: string;
-  templateKey: string;
-  name: string;
-  description: string;
-}
-
-export const DEFAULT_STYLE_TEMPLATES: DefaultTemplateDefinition[] = [
-  {
-    key: "bottom-loop-reality",
-    name: "底层循环现实流",
-    description: "通过碎片化生活与反复落空表现人物困境。",
-    category: "现实流",
-    tags: ["第一人称", "口语化", "碎片叙事"],
-    applicableGenres: ["都市", "现实", "成长"],
-    analysisMarkdown: "以时间推进和现实落差构成叙事张力，结尾不解决核心困境。",
-    narrativeRules: {
-      progressionMode: "time_sequence",
-      sceneUnitPattern: ["行为", "落差", "自我合理化"],
-      multiPov: false,
-      looping: true,
-      endingStyle: "unresolved",
-      summary: "以碎片化生活推进，不做总括式回顾。",
-    },
-    characterRules: {
-      allowSelfReflection: false,
-      emotionExpression: "behavior_only",
-      defenseMechanisms: ["嘴硬", "转移", "自我合理化"],
-      facePriority: true,
-      dialogueStyle: "short_colloquial",
-      summary: "人物情绪通过动作和嘴硬表达。",
-    },
-    languageRules: {
-      register: "colloquial",
-      roughness: 0.8,
-      allowIncompleteSentences: true,
-      allowSwearing: true,
-      sentenceVariation: "high",
-      allowUselessDetails: true,
-      summary: "语言粗粝、口语化，允许生活杂音。",
-    },
-    rhythmRules: {
-      pace: "medium_fast",
-      paragraphDensity: "high",
-      allowFragmentedFlow: true,
-      actionOverExplanation: true,
-      summary: "段落密实，动作先于解释。",
-    },
-    defaultAntiAiRuleKeys: [
-      "forbid-explicit-psychology",
-      "forbid-ending-elevation",
-      "encourage-useless-action",
-      "encourage-reality-gap",
-      "encourage-hard-mouth-compensation",
-    ],
-  },
-  {
-    key: "power-up-escalation",
-    name: "爽文递进推进流",
-    description: "持续升级冲突和收益点，强化目标推进与爽点兑现。",
-    category: "爽文流",
-    tags: ["推进感", "收益点", "冲突升级"],
-    applicableGenres: ["都市", "玄幻", "热血"],
-    analysisMarkdown: "每段都要有目标推进或爽点兑现，保持明确因果和节奏抬升。",
-    narrativeRules: {
-      progressionMode: "goal_driven",
-      sceneUnitPattern: ["目标", "阻碍", "压制", "反转收益"],
-      multiPov: false,
-      looping: false,
-      endingStyle: "hook",
-      summary: "围绕目标推进，尽快兑现局部收益。",
-    },
-    characterRules: {
-      allowSelfReflection: true,
-      emotionExpression: "dialogue_and_action",
-      defenseMechanisms: [],
-      facePriority: false,
-      dialogueStyle: "direct",
-      summary: "角色表达直接，情绪跟随胜负切换。",
-    },
-    languageRules: {
-      register: "direct",
-      roughness: 0.55,
-      allowIncompleteSentences: false,
-      allowSwearing: false,
-      sentenceVariation: "medium",
-      allowUselessDetails: false,
-      summary: "句式清晰，减少无效分散信息。",
-    },
-    rhythmRules: {
-      pace: "fast",
-      paragraphDensity: "medium",
-      allowFragmentedFlow: false,
-      actionOverExplanation: true,
-      summary: "优先冲突和结果，少停留。",
-    },
-    defaultAntiAiRuleKeys: [
-      "forbid-theme-summary",
-      "risk-dialogue-too-functional",
-      "risk-three-paragraphs-exposition",
-    ],
-  },
-  {
-    key: "suspense-pressure",
-    name: "悬疑压迫递增流",
-    description: "通过信息遮蔽、细节异常和压力叠加制造不安感。",
-    category: "悬疑流",
-    tags: ["压迫感", "信息差", "异常细节"],
-    applicableGenres: ["悬疑", "惊悚", "现实"],
-    analysisMarkdown: "以异常细节、信息差和节奏收束推动悬念层层加压。",
-    narrativeRules: {
-      progressionMode: "mystery_escalation",
-      sceneUnitPattern: ["现场细节", "异常", "误判", "新风险"],
-      multiPov: false,
-      looping: false,
-      endingStyle: "suspense",
-      summary: "优先制造信息缺口和压迫氛围。",
-    },
-    characterRules: {
-      allowSelfReflection: true,
-      emotionExpression: "reaction_only",
-      defenseMechanisms: ["压抑"],
-      facePriority: false,
-      dialogueStyle: "restrained",
-      summary: "角色反应克制，恐惧通过反应显现。",
-    },
-    languageRules: {
-      register: "restrained",
-      roughness: 0.45,
-      allowIncompleteSentences: true,
-      allowSwearing: false,
-      sentenceVariation: "medium_high",
-      allowUselessDetails: true,
-      summary: "细节精确，保留少量噪音增强现场感。",
-    },
-    rhythmRules: {
-      pace: "medium",
-      paragraphDensity: "medium_high",
-      allowFragmentedFlow: true,
-      actionOverExplanation: true,
-      summary: "通过节奏收束和信息延迟制造压力。",
-    },
-    defaultAntiAiRuleKeys: [
-      "forbid-explicit-psychology",
-      "forbid-theme-summary",
-      "risk-even-paragraph-length",
-      "encourage-reality-gap",
-    ],
-  },
-  {
-    key: "emotional-tension",
-    name: "情绪拉扯流",
-    description: "通过错位表达、停顿和误读制造关系张力。",
-    category: "情感流",
-    tags: ["误读", "拉扯", "停顿感"],
-    applicableGenres: ["言情", "都市", "群像"],
-    analysisMarkdown: "人物不直说核心情绪，靠误读、停顿和反应推动关系变化。",
-    narrativeRules: {
-      progressionMode: "relationship_push_pull",
-      sceneUnitPattern: ["动作", "言外之意", "误读", "回避"],
-      multiPov: false,
-      looping: false,
-      endingStyle: "emotional_hook",
-      summary: "以关系错位推进，而非直接说明。",
-    },
-    characterRules: {
-      allowSelfReflection: true,
-      emotionExpression: "subtext",
-      defenseMechanisms: ["回避", "试探", "嘴硬"],
-      facePriority: true,
-      dialogueStyle: "subtext_heavy",
-      summary: "情绪通过停顿、动作和言外之意体现。",
-    },
-    languageRules: {
-      register: "natural",
-      roughness: 0.35,
-      allowIncompleteSentences: true,
-      allowSwearing: false,
-      sentenceVariation: "high",
-      allowUselessDetails: true,
-      summary: "语言自然，允许留白与停顿。",
-    },
-    rhythmRules: {
-      pace: "medium_slow",
-      paragraphDensity: "medium",
-      allowFragmentedFlow: true,
-      actionOverExplanation: false,
-      summary: "给关系反应留空间，但避免空洞抒情。",
-    },
-    defaultAntiAiRuleKeys: [
-      "forbid-direct-preaching",
-      "forbid-ending-elevation",
-      "risk-dialogue-too-functional",
-      "encourage-useless-action",
-    ],
-  },
-  {
-    key: "ensemble-weave",
-    name: "群像交织流",
-    description: "以多人行动线和视角差异交织推进事件。",
-    category: "群像流",
-    tags: ["多角色", "交织", "信息流动"],
-    applicableGenres: ["群像", "都市", "悬疑"],
-    analysisMarkdown: "多角色并行推进，但每个角色的表达和认知范围必须区分清楚。",
-    narrativeRules: {
-      progressionMode: "multi_thread",
-      sceneUnitPattern: ["角色动作", "局部信息", "交叉影响"],
-      multiPov: true,
-      looping: false,
-      endingStyle: "cross_hook",
-      povSwitchStyle: "controlled",
-      summary: "多线并进，但视角切换要受控。",
-    },
-    characterRules: {
-      allowSelfReflection: true,
-      emotionExpression: "mixed",
-      defenseMechanisms: [],
-      facePriority: false,
-      dialogueStyle: "distinct_by_role",
-      summary: "不同角色口吻必须拉开差异。",
-    },
-    languageRules: {
-      register: "flexible",
-      roughness: 0.45,
-      allowIncompleteSentences: true,
-      allowSwearing: false,
-      sentenceVariation: "high",
-      allowUselessDetails: false,
-      summary: "保持角色差异，避免所有人说话一样。",
-    },
-    rhythmRules: {
-      pace: "balanced",
-      paragraphDensity: "medium",
-      allowFragmentedFlow: false,
-      actionOverExplanation: true,
-      summary: "多线交织但节奏不乱。",
-    },
-    defaultAntiAiRuleKeys: [
-      "risk-dialogue-too-functional",
-      "risk-repeated-sentence-structure",
-      "forbid-theme-summary",
-    ],
-  },
-  {
-    key: "immersive-daily",
-    name: "日常浸没流",
-    description: "通过生活细节和细微情绪变化建立持续沉浸感。",
-    category: "日常流",
-    tags: ["生活感", "沉浸", "细碎细节"],
-    applicableGenres: ["日常", "治愈", "都市"],
-    analysisMarkdown: "允许保留生活性动作和无效信息，但核心情绪仍要通过场景自然流出。",
-    narrativeRules: {
-      progressionMode: "scene_immersion",
-      sceneUnitPattern: ["动作", "环境", "关系反应"],
-      multiPov: false,
-      looping: false,
-      endingStyle: "soft_open",
-      summary: "重场景体验和关系温度。",
-    },
-    characterRules: {
-      allowSelfReflection: true,
-      emotionExpression: "light_behavior",
-      defenseMechanisms: [],
-      facePriority: false,
-      dialogueStyle: "daily_natural",
-      summary: "人物表达自然，不用高强度戏剧句。",
-    },
-    languageRules: {
-      register: "colloquial",
-      roughness: 0.25,
-      allowIncompleteSentences: true,
-      allowSwearing: false,
-      sentenceVariation: "medium_high",
-      allowUselessDetails: true,
-      summary: "保留生活细节和杂音，不追求工整。",
-    },
-    rhythmRules: {
-      pace: "slow",
-      paragraphDensity: "medium",
-      allowFragmentedFlow: true,
-      actionOverExplanation: false,
-      summary: "慢节奏沉浸，但避免空转。",
-    },
-    defaultAntiAiRuleKeys: [
-      "forbid-ending-elevation",
-      "risk-even-paragraph-length",
-      "encourage-useless-action",
-    ],
-  },
-  {
-    key: "cold-professional",
-    name: "冷峻专业流",
-    description: "以专业事实和行业细节压住情绪，形成克制压力感。",
-    category: "专业流",
-    tags: ["专业细节", "克制", "事实压情绪"],
-    applicableGenres: ["职场", "现实", "悬疑"],
-    analysisMarkdown: "行业事实和程序细节优先，情绪不直说，信息密度高于抒情密度。",
-    narrativeRules: {
-      progressionMode: "fact_driven",
-      sceneUnitPattern: ["事实", "动作", "专业判断", "后果"],
-      multiPov: false,
-      looping: false,
-      endingStyle: "pressure_continue",
-      summary: "让专业事实承担叙事重量。",
-    },
-    characterRules: {
-      allowSelfReflection: false,
-      emotionExpression: "suppressed",
-      defenseMechanisms: ["克制"],
-      facePriority: false,
-      dialogueStyle: "informational",
-      summary: "情绪藏在专业动作和事实选择里。",
-    },
-    languageRules: {
-      register: "professional",
-      roughness: 0.2,
-      allowIncompleteSentences: false,
-      allowSwearing: false,
-      sentenceVariation: "medium",
-      allowUselessDetails: false,
-      summary: "术语和事实优先，避免廉价金句。",
-    },
-    rhythmRules: {
-      pace: "balanced",
-      paragraphDensity: "medium_high",
-      allowFragmentedFlow: false,
-      actionOverExplanation: true,
-      summary: "信息密度高，但不铺张解释。",
-    },
-    defaultAntiAiRuleKeys: [
-      "forbid-direct-preaching",
-      "forbid-theme-summary",
-      "risk-repeated-sentence-structure",
-    ],
-  },
-  {
-    key: "absurd-dark-humor",
-    name: "荒诞黑色幽默流",
-    description: "通过反差、冷感观察和荒诞细节制造黑色幽默。",
-    category: "黑色幽默",
-    tags: ["荒诞", "反差", "冷感"],
-    applicableGenres: ["都市", "黑色幽默", "现实"],
-    analysisMarkdown: "用反差和荒诞细节放大现实困境，笑点和压迫感同时存在。",
-    narrativeRules: {
-      progressionMode: "contrast_driven",
-      sceneUnitPattern: ["现实细节", "荒诞偏差", "冷反应"],
-      multiPov: false,
-      looping: false,
-      endingStyle: "bitter_aftertaste",
-      summary: "依赖反差和冷感观察，而非热闹吐槽。",
-    },
-    characterRules: {
-      allowSelfReflection: false,
-      emotionExpression: "deadpan",
-      defenseMechanisms: ["自嘲", "转移"],
-      facePriority: true,
-      dialogueStyle: "deadpan_colloquial",
-      summary: "情绪藏在冷反应和嘴硬里。",
-    },
-    languageRules: {
-      register: "colloquial",
-      roughness: 0.5,
-      allowIncompleteSentences: true,
-      allowSwearing: true,
-      sentenceVariation: "high",
-      allowUselessDetails: true,
-      summary: "允许夹带荒诞杂质和冷幽默节奏。",
-    },
-    rhythmRules: {
-      pace: "balanced",
-      paragraphDensity: "medium_high",
-      allowFragmentedFlow: true,
-      actionOverExplanation: true,
-      summary: "反差点要快落地，不要解释笑点。",
-    },
-    defaultAntiAiRuleKeys: [
-      "forbid-explicit-psychology",
-      "forbid-ending-elevation",
-      "encourage-reality-gap",
-      "encourage-hard-mouth-compensation",
-    ],
-  },
-  {
-    key: "webnovel-human-continuity",
-    name: "网文自然人味连续流",
-    description: "默认给长篇网文使用：弱化 AI 腔，强化现场推进、人物口吻和连续性。",
-    category: "长篇网文",
-    tags: ["去AI味", "连续性", "人味", "网文"],
-    applicableGenres: ["玄幻", "都市", "悬疑", "科幻", "言情", "历史"],
-    analysisMarkdown: "先保证每章像真实现场，而不是像模型在总结设定：用动作、对话、代价和章末压力承接长篇连续性。",
-    narrativeRules: {
-      progressionMode: "chapter_chain",
-      sceneUnitPattern: ["当前压力", "角色动作", "信息变化", "章末钩子"],
-      multiPov: false,
-      looping: false,
-      endingStyle: "hook",
-      summary: "每章只解决局部问题，同时留下新的压力、选择或误解。",
-    },
-    characterRules: {
-      allowSelfReflection: false,
-      emotionExpression: "dialogue_and_action",
-      defenseMechanisms: ["嘴硬", "试探", "遮掩", "找补"],
-      facePriority: true,
-      dialogueStyle: "distinct_by_role",
-      summary: "人物通过动作、口头禅、回避和选择暴露状态，不替他们解释。",
-    },
-    languageRules: {
-      register: "colloquial",
-      roughness: 0.55,
-      allowIncompleteSentences: true,
-      allowSwearing: false,
-      sentenceVariation: "high",
-      allowUselessDetails: true,
-      summary: "语言要像人在现场说话和观察，保留不工整的短句、停顿和生活杂音。",
-    },
-    rhythmRules: {
-      pace: "medium_fast",
-      paragraphDensity: "medium_high",
-      allowFragmentedFlow: true,
-      actionOverExplanation: true,
-      summary: "先让事情发生，再让读者从后果里理解。",
-    },
-    defaultAntiAiRuleKeys: [
-      "forbid-template-transitions",
-      "forbid-generic-metaphor",
-      "forbid-setting-info-dump",
-      "forbid-explicit-psychology",
-      "forbid-theme-summary",
-      "risk-dialogue-too-functional",
-      "risk-repeated-sentence-structure",
-      "risk-even-paragraph-length",
-      "encourage-useless-action",
-      "encourage-life-noise",
-      "encourage-character-specific-voice",
-    ],
-  },
-];
-
-export const DEFAULT_STARTER_STYLE_PROFILES: DefaultStarterStyleProfileDefinition[] = [
-  {
-    key: "starter-power-up",
-    templateKey: "power-up-escalation",
-    name: "我的默认爽文推进写法",
-    description: "适合第一次开书先跑顺目标推进、爽点兑现和章节收益点，后续可以直接在此基础上微调。",
-  },
-  {
-    key: "starter-suspense",
-    templateKey: "suspense-pressure",
-    name: "我的默认悬疑压迫写法",
-    description: "适合异常、规则、调查和危险逼近类故事，先帮你把压迫感和信息差稳住。",
-  },
-  {
-    key: "starter-emotional",
-    templateKey: "emotional-tension",
-    name: "我的默认情绪拉扯写法",
-    description: "适合关系推进、误读拉扯和情绪兑现类故事，先有一套能直接开写的关系型表达底座。",
-  },
-  {
-    key: "starter-daily",
-    templateKey: "immersive-daily",
-    name: "我的默认日常浸没写法",
-    description: "适合治愈、陪伴、生活经营和轻缓成长类故事，优先保证生活感和沉浸感。",
-  },
-  {
-    key: "starter-webnovel-human",
-    templateKey: "webnovel-human-continuity",
-    name: "我的默认网文去 AI 味写法",
-    description: "适合作为长篇网文默认底座：用户只做方向判断，AI 负责规划、写章、审校、回灌事实。",
-  },
-];
-
-export const DEFAULT_ANTI_AI_RULES: DefaultAntiAiRuleDefinition[] = [
-  {
-    key: "forbid-explicit-psychology",
-    name: "禁止解释型心理描写",
-    type: "forbidden",
-    severity: "high",
-    description: "禁止直接使用“他感到”“他意识到”等句式解释人物心理。",
-    detectPatterns: ["他感到", "她感到", "他意识到", "她意识到", "他明白了", "她明白了"],
-    rewriteSuggestion: "把心理解释改成动作、语气、停顿、环境反应或结果。",
-    promptInstruction: "禁止直接解释人物心理，人物状态必须通过行为、动作、语气或环境反应体现。",
-    autoRewrite: true,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "forbid-ending-elevation",
-    name: "禁止段尾升华",
-    type: "forbidden",
-    severity: "high",
-    description: "禁止在段尾或收尾处用总结句升华主题。",
-    detectPatterns: ["生活就是", "命运总会", "归根结底", "说到底", "这就是"],
-    rewriteSuggestion: "删除升华句，回到具体动作、现场或悬而未决的处境。",
-    promptInstruction: "禁止段尾升华或总结主题，收尾必须落回具体情境。",
-    autoRewrite: true,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "forbid-theme-summary",
-    name: "禁止总结主题",
-    type: "forbidden",
-    severity: "high",
-    description: "禁止把段落写成总结中心思想或提炼人生道理。",
-    detectPatterns: ["这说明", "这意味着", "归根结底", "其实就是"],
-    rewriteSuggestion: "删掉主题总结，让信息通过事件和结果自然显现。",
-    promptInstruction: "不要总结主题，不要替读者提炼中心思想。",
-    autoRewrite: true,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "forbid-direct-preaching",
-    name: "禁止直接说教",
-    type: "forbidden",
-    severity: "medium",
-    description: "禁止作者替角色或读者做直接价值判断和说教。",
-    detectPatterns: ["我们都应该", "人总要学会", "真正重要的是"],
-    rewriteSuggestion: "改成角色具体处境或对话，不做抽象说教。",
-    promptInstruction: "禁止直接说教或价值判断，让意义落在事件本身。",
-    autoRewrite: true,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "forbid-template-transitions",
-    name: "禁止模板化转场和导游词",
-    type: "forbidden",
-    severity: "high",
-    description: "禁止用“这一刻”“与此同时”“接下来”等提示词式转场或结论句推动章节。",
-    detectPatterns: ["这一刻", "与此同时", "接下来", "他知道，", "她知道，", "毫无疑问", "不难看出"],
-    rewriteSuggestion: "改用角色动作、视线变化、声音、物件或一句有信息量的对话完成场景切换。",
-    promptInstruction: "禁止写作提示词式转场、结论式提示和段首导游词；转场必须落在动作、视线、声音、物件或对话上。",
-    autoRewrite: true,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "forbid-generic-metaphor",
-    name: "禁止通用 AI 腔比喻和泛化形容",
-    type: "forbidden",
-    severity: "high",
-    description: "禁止使用“仿佛”“难以言喻”“复杂情绪”等可套进任何场景的泛化表达。",
-    detectPatterns: ["仿佛", "像是某种", "空气仿佛凝固", "命运的齿轮", "某种意义上", "无形之中", "复杂的情绪", "难以言喻"],
-    rewriteSuggestion: "替换为当前角色、地点、物件和冲突独有的具体细节。",
-    promptInstruction: "禁止通用 AI 腔套话和泛化形容词；必须换成当前角色或场景独有的具体细节。",
-    autoRewrite: true,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "forbid-setting-info-dump",
-    name: "禁止说明书式设定倾倒",
-    type: "forbidden",
-    severity: "high",
-    description: "禁止连续整段介绍世界规则、修炼体系、组织结构或背景资料。",
-    detectPatterns: ["这个世界", "所谓", "规则是", "体系分为", "简单来说", "众所周知"],
-    rewriteSuggestion: "把设定拆进冲突、选择、代价、误会、交易或失败后果里释放。",
-    promptInstruction: "设定只能在冲突、选择、代价、对话误解或后果中释放，不得写成说明书。",
-    autoRewrite: true,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "risk-even-paragraph-length",
-    name: "段落长度过于整齐",
-    type: "risk",
-    severity: "medium",
-    description: "段落长度和节奏过于平均，容易产生 AI 作文感。",
-    detectPatterns: ["段落整齐", "节奏平均"],
-    rewriteSuggestion: "打破段落长度均衡，让句子和段落有自然起伏。",
-    promptInstruction: "注意避免每段都过于工整、平均和像标准作文。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "risk-three-paragraphs-exposition",
-    name: "连续三段解释性叙事",
-    type: "risk",
-    severity: "high",
-    description: "连续几段只有解释没有动作，会削弱现场感。",
-    detectPatterns: ["连续解释", "没有动作"],
-    rewriteSuggestion: "插入动作、对话、环境反馈，减少连段说明。",
-    promptInstruction: "注意避免连续几段只有解释没有动作或对话。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "risk-dialogue-too-functional",
-    name: "对话纯功能推进",
-    type: "risk",
-    severity: "medium",
-    description: "对话只有信息推进，没有人物语气和生活噪音。",
-    detectPatterns: ["告诉你", "我们现在要", "接下来就"],
-    rewriteSuggestion: "补入停顿、绕弯、语气差异和无效信息。",
-    promptInstruction: "对话不要只承担推进剧情功能，要保留人物语气和生活噪音。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "risk-repeated-sentence-structure",
-    name: "句式重复率过高",
-    type: "risk",
-    severity: "medium",
-    description: "连续句式过于整齐，容易显得机械。",
-    detectPatterns: ["首先", "然后", "接着", "最后"],
-    rewriteSuggestion: "拉开句式长度和起句方式，打散结构。",
-    promptInstruction: "警惕连续句式重复和排比化表达。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: true,
-  },
-  {
-    key: "encourage-useless-action",
-    name: "鼓励无意义小动作",
-    type: "encourage",
-    severity: "low",
-    description: "鼓励加入真实但不推动主线的小动作增强人味。",
-    detectPatterns: [],
-    rewriteSuggestion: "补入挠头、点烟、抠包装、挪椅子等小动作。",
-    promptInstruction: "优先加入真实但不推动主线的小动作，增加生活感。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: false,
-  },
-  {
-    key: "encourage-reality-gap",
-    name: "鼓励现实落差",
-    type: "encourage",
-    severity: "low",
-    description: "让人物预期和现实结果之间出现落差。",
-    detectPatterns: [],
-    rewriteSuggestion: "补出人物预期与实际结果之间的差距。",
-    promptInstruction: "优先制造现实落差，让人物预期和结果不完全一致。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: false,
-  },
-  {
-    key: "encourage-hard-mouth-compensation",
-    name: "鼓励嘴硬补偿",
-    type: "encourage",
-    severity: "low",
-    description: "人物吃瘪后用嘴硬、转移或找补来维持体面。",
-    detectPatterns: [],
-    rewriteSuggestion: "给角色补一句嘴硬找补或自我合理化。",
-    promptInstruction: "人物吃瘪后可用嘴硬、转移或找补来补偿体面。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: false,
-  },
-  {
-    key: "encourage-life-noise",
-    name: "鼓励生活噪音",
-    type: "encourage",
-    severity: "low",
-    description: "适当加入与主线不完全相关的生活噪音增加现场感。",
-    detectPatterns: [],
-    rewriteSuggestion: "补入环境声、杂音、小打断或生活性信息。",
-    promptInstruction: "适当保留生活噪音和无效信息，增强现场感。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: false,
-  },
-  {
-    key: "encourage-character-specific-voice",
-    name: "鼓励角色专属口吻",
-    type: "encourage",
-    severity: "low",
-    description: "重要角色需要有可辨认的说话节奏、避讳、误解方式和小动作。",
-    detectPatterns: [],
-    rewriteSuggestion: "为每个重要角色补入口头节奏、避讳词、习惯动作或只属于他的误读方式。",
-    promptInstruction: "重要角色必须有可识别的口头节奏、避讳、误解方式和小动作；对话不能像同一个旁白代言。",
-    autoRewrite: false,
-    enabled: true,
-    globalBaselineEnabled: false,
-  },
-];
+export type {
+  DefaultAntiAiRuleDefinition,
+  DefaultStarterStyleProfileDefinition,
+  DefaultTemplateDefinition,
+} from "./defaultTypes";
+export { DEFAULT_STYLE_TEMPLATES } from "./defaultStyleTemplates";
+export { DEFAULT_STARTER_STYLE_PROFILES } from "./defaultStarterStyleProfiles";
+export { DEFAULT_ANTI_AI_RULES } from "./defaultAntiAiRules";
